% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lsg.R
\name{make_lsg}
\alias{make_lsg}
\title{Calculate LSG}
\usage{
make_lsg(
  dataset = dataset,
  crit_to_4plus = NULL,
  crit_to_4 = NULL,
  crit_to_3 = NULL,
  non_crit = NULL
)
}
\arguments{
\item{dataset}{Dataset you want to use to calculate the LSG}

\item{crit_to_4plus}{Vector with the names of the critical indicators coded from 1 to 5.}

\item{crit_to_4}{Vector with the names of the critical indicators coded from 1 to 4.}

\item{crit_to_3}{Vector with the names of the critical indicators coded from 1 to 3.}

\item{non_crit}{Name of the non-critical indicator}
}
\value{
A numeric vector with the LSG score from 1 to 5 (or maximum).
}
\description{
Calculate LSG
}
\examples{
make_lsg(dummy_dataset,
         crit_to_4plus = c("crit1_4plus", "crit2_4plus"),
         crit_to_4 = c("crit1_4", "crit2_4"),
         crit_to_3 = c("crit1_3", "crit2_3"),
         non_crit = "crit1_nc")

}
