% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lsg.R
\name{make_single_lsg}
\alias{make_single_lsg}
\title{Calculate LSG. This function is made to be used in the make_lsg.}
\usage{
make_single_lsg(dataset, max_indicator, indicator_to_select)
}
\arguments{
\item{dataset}{Dataset you want to use to calculate the LSG}

\item{max_indicator}{Vector with all the names of the variables with the maximum score}

\item{indicator_to_select}{Vector with all the names of the variables that compose the LSG (score from 1 to 5)}
}
\value{
A numeric vector with the LSG score 1 to the maximum score.
}
\description{
Calculate LSG. This function is made to be used in the make_lsg.
}
\examples{
make_single_lsg(dataset = dummy_dataset,
                max_indicator = c("crit1_4plus", "crit2_4plus"),
                indicator_to_select = c("crit1_4plus", "crit2_4plus", "crit1_4", "crit2_4", "crit1_3", "crit2_3", "crit1_nc"))

}
